+++ aws/eks-cluster/datadog-rbac.tf
@@ -3,6 +3,7 @@ module "datadog_rbac" {
   source                          = "./modules/datadog-rbac"
   fargate_profiles                = var.fargate_profiles
   default_service_account         = var.default_service_account
+  custom_service_accounts         = var.custom_service_accounts
   custom_namespaces               = var.custom_namespaces
   datadog_agent_cluster_role_name = var.datadog_agent_cluster_role_name
 }